Modifications pour faire fonctionner les tests

+++ GROUND.PLATEFORM/NUnittest/DataPackageTests/IntegrationTestsBase.cs
@@ -32,6 +32,7 @@ using PIS.Ground.DataPackage;
 using PIS.Ground.DataPackage.RemoteDataStoreFactory;
 using PIS.Ground.DataPackage.RequestMgt;
 using PIS.Ground.GroundCore.AppGround;
+using System.Windows.Forms;
 using CommLinkEnum = DataPackageTests.T2GServiceInterface.Identification.commLinkEnum;
 
 namespace DataPackageTests
@@ -489,12 +490,15 @@ namespace DataPackageTests
 
                 if (File.Exists(_databaseFilePath))
                 {
-                    string cmdDropDB =
-                        "IF EXISTS( select name from sys.databases where NAME= '" + _databaseName + "')" +
-                        " BEGIN DROP DATABASE [" + _databaseName + "] END";
+                    bool test = false;
+                    if(!test) {
+                        string cmdDropDB =
+                            "IF EXISTS( select name from sys.databases where NAME= '" + _databaseName + "')" +
+                            " BEGIN DROP DATABASE [" + _databaseName + "] END";
 
-                    SqlHelper.ExecuteNonQuery(HistoryLoggerConfiguration.SqlCreateDbConnectionString, System.Data.CommandType.Text, cmdDropDB);
-                    File.Delete(_databaseFilePath);
+                        SqlHelper.ExecuteNonQuery(HistoryLoggerConfiguration.SqlCreateDbConnectionString, System.Data.CommandType.Text, cmdDropDB);
+                        File.Delete(_databaseFilePath);
+                    }
                 }
 
                 if (File.Exists(_databaseLogPath))
